Restated `freshness_theorem_some_any`
In `theories/FreshnessTheorem.v`, the lemma `freshness_theorem_some_any`
was simplified, as the equivalence stated in there wasn't needed in
future proofs. As one side of the equivalence is assumed in the context
assumptions, the statement's position might be unintuitive or confusing,
as one direction might be proven directly using the assumption `Hp`.

Alternative idea, if one wants to keep its generality: Prove the lemma
before introducing the assumption `Hp`.